# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Webstore Customer API",
        "description": "Dokumentasi endpoint API customer untuk autentikasi dan fitur terkait dashboard.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Default API server"
        }
    ],
    "paths": {
...
        "/api/dashboard": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Data dashboard customer",
                "description": "Mengembalikan seluruh data dashboard customer seperti profil, order, wallet, bonus, dan jaringan.",
                "operationId": "d5fcadff7c52501c317a59bc00886f22",
                "parameters": [
                    {
                        "name": "orders_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "wallet_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "wallet_search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "wallet_type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "wallet_status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data dashboard berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Data dashboard berhasil diambil.",
                                    "data": {
                                        "customer": {
                                            "id": 1,
                                            "name": "Budi Santoso"
                                        },
                                        "stats": {
                                            "orders_total": 12,
                                            "wallet_balance": 150000
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Tidak terautentikasi."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/dashboard/leaderboards": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Leaderboard customer",
                "description": "Mengembalikan data leaderboard customer dengan filter periode harian, mingguan, dan bulanan.",
                "operationId": "c8c84ae051babd7ff3ca89234973b440",
                "parameters": [
                    {
                        "name": "tab",
                        "in": "query",
                        "description": "Filter tab: 1=Harian, 2=Mingguan, 3=Bulanan",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                1,
                                2,
                                3
                            ]
                        }
                    },
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Alternatif filter periode: daily/weekly/monthly atau harian/mingguan/bulanan",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "daily",
                                "weekly",
                                "monthly",
                                "harian",
                                "mingguan",
                                "bulanan"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Leaderboard berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Leaderboard fetched successfully",
                                    "data": {
                                        "tabs": [
                                            "Harian",
                                            "Mingguan",
                                            "Bulanan"
                                        ],
                                        "selected_tab": 1,
                                        "my_rank": {
                                            "rank": 12,
                                            "name": "Budi Santoso",
                                            "avatar": "BS",
                                            "level": "Gold Member",
                                            "trend": "up",
                                            "streak": 7,
                                            "points": 18500
                                        },
                                        "leaderboard": [
                                            {
                                                "id": 101,
                                                "name": "Andi Wijaya",
                                                "avatar": "AW",
                                                "level": "Diamond",
                                                "trend": "up",
                                                "streak": 21,
                                                "points": 125000
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Tidak terautentikasi."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "The given data was invalid."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token dengan format: Bearer {token}",
                "bearerFormat": "Bearer Token",
                "scheme": "bearer"
            }
        }
    },
...
}